documentationtools.compare_images() now returns 1 if ImageMagick's ``compare`` is not available.

+++ trunk/abjad/tools/documentationtools/compare_images.py
@@ -9,29 +9,37 @@ def compare_images(image_one, image_two):
     commandline tool.
 
     Return `True` if images are the same, otherwise `False`.
+
+    If `compare` is not available, return 1.
     '''
+    from abjad.tools import iotools
 
     assert os.path.exists(image_one)
     assert os.path.exists(image_two)
 
-    tempdir = tempfile.mkdtemp()
-    comparison = os.path.join(tempdir, 'comparison.png')
+    if iotools.which('compare'):
 
-    command = 'compare -metric ae {} {} {}'.format(
-        image_one, image_two, comparison)
-    process = subprocess.Popen(command.split(),
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE
-        )
+        tempdir = tempfile.mkdtemp()
+        comparison = os.path.join(tempdir, 'comparison.png')
 
-    stderr = process.stderr.read()
-    result = True
-    if stderr.startswith('compare: image widths or heights differ'):
-        result = False
-    else:
-        # FIXME: following line always errors
-        #result = int(stderr.split('\n')[0].split()[0]) is 0
-        pass
+        command = 'compare -metric ae {} {} {}'.format(
+            image_one, image_two, comparison)
+        process = subprocess.Popen(command.split(),
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE
+            )
 
-    shutil.rmtree(tempdir)
-    return result
+        stderr = process.stderr.read()
+        result = True
+        if stderr.startswith('compare: image widths or heights differ'):
+            result = False
+        else:
+            # FIXME: following line always errors
+            result = int(stderr.split('\n')[0].split()[0]) is 0
+            #pass
+
+        shutil.rmtree(tempdir)
+
+        return result
+
+    return 1
